removed undeployed version of federation contracts
